backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
discord.py==2.3.2
openai==1.52.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pillow==10.1.0
numpy==1.24.4
scikit-learn==1.3.0
aiofiles==23.2.1
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2
transformers>=4.21.0
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0

email-validator>=2.0.0